deprecate redundant frequency_band in lieu of band
- merge functionality of frequency_band.py into band.py
