+++ mt_metadata/transfer_functions/processing/aurora/__init__.py
@@ -5,7 +5,6 @@ from .window import Window
 from .decimation import Decimation
 from .decimation_level import DecimationLevel
 from .estimator import Estimator
-from .frequency_band import FrequencyBand
 from .processing import Processing
 from .regression import Regression
 from .run import Run
