+++ mt_metadata/transfer_functions/processing/aurora/decimation_level.py
@@ -8,20 +8,17 @@ Created on Thu Feb 17 14:15:20 2022
 # Imports
 # =============================================================================
 import numpy as np
+import pandas as pd
 
 from mt_metadata.base.helpers import write_lines
 from mt_metadata.base import get_schema, Base
-from mt_metadata.transfer_functions.processing.aurora.frequency_band import (
-    get_fft_harmonics,
-)
 
-from .standards import SCHEMA_FN_PATHS
-
-from .window import Window
-from .decimation import Decimation
 from .band import Band
-from .regression import Regression
+from .decimation import Decimation
 from .estimator import Estimator
+from .regression import Regression
+from .standards import SCHEMA_FN_PATHS
+from .window import Window
 
 # =============================================================================
 attr_dict = get_schema("decimation_level", SCHEMA_FN_PATHS)
@@ -32,6 +29,78 @@ attr_dict.add_dict(get_schema("estimator", SCHEMA_FN_PATHS), "estimator")
 
 
 # =============================================================================
+
+
+def df_from_bands(band_list):
+    """
+    Utility function that transforms a list of bands into a dataframe
+
+    Note: The decimation_level here is +1 to agree with EMTF convention.
+        Not clear this is really necessary
+
+    Parameters
+    ----------
+    band_list: list
+        obtained from mt_metadata.transfer_functions.processing.aurora.decimation_level.DecimationLevel.bands
+
+    Returns
+    -------
+    out_df: pd.Dataframe
+        Same format as that generated by EMTFBandSetupFile.get_decimation_level()
+    """
+    df_columns = [
+        "decimation_level",
+        "lower_bound_index",
+        "upper_bound_index",
+        "frequency_min",
+        "frequency_max",
+    ]
+    n_rows = len(band_list)
+    df_columns_dict = {}
+    for col in df_columns:
+        df_columns_dict[col] = n_rows * [None]
+    for i_band, band in enumerate(band_list):
+        df_columns_dict["decimation_level"][i_band] = band.decimation_level + 1
+        df_columns_dict["lower_bound_index"][i_band] = band.index_min
+        df_columns_dict["upper_bound_index"][i_band] = band.index_max
+        df_columns_dict["frequency_min"][i_band] = band.frequency_min
+        df_columns_dict["frequency_max"][i_band] = band.frequency_max
+    out_df = pd.DataFrame(data=df_columns_dict)
+    out_df.sort_values(by="lower_bound_index", inplace=True)
+    out_df.reset_index(inplace=True, drop=True)
+    return out_df
+
+def get_fft_harmonics(samples_per_window, sample_rate):
+    """
+    Works for odd and even number of points.
+
+    Could be midified with kwargs to support one_sided, two_sided, ignore_dc
+    ignore_nyquist, and etc.  Could actally take FrequencyBands as an argument
+    if we wanted as well.
+
+    Parameters
+    ----------
+    samples_per_window: integer
+        Number of samples in a window that will be Fourier transformed.
+    sample_rate: float
+            Inverse of time step between samples,
+            Samples per second
+
+    Returns
+    -------
+    harmonic_frequencies: numpy array
+        The frequencies that the fft will be computed.
+        These are one-sided (positive frequencies only)
+        Does not return Nyquist
+        Does return DC component
+    """
+    n_fft_harmonics = int(samples_per_window / 2)  # no bin at Nyquist,
+    delta_t = 1.0 / sample_rate
+    harmonic_frequencies = np.fft.fftfreq(samples_per_window, d=delta_t)
+    harmonic_frequencies = harmonic_frequencies[0:n_fft_harmonics]
+    return harmonic_frequencies
+
+
 class DecimationLevel(Base):
     __doc__ = write_lines(attr_dict)
 
@@ -145,10 +214,6 @@ class DecimationLevel(Base):
         bands_df: pd.Dataframe
             Same format as that generated by EMTFBandSetupFile.get_decimation_level()
         """
-        from mt_metadata.transfer_functions.processing.aurora.frequency_band import (
-            df_from_bands,
-        )
-
         bands_df = df_from_bands(self.bands)
         return bands_df
 
@@ -174,7 +239,7 @@ class DecimationLevel(Base):
         -------
 
         """
-        from mt_metadata.transfer_functions.processing.aurora.frequency_band import (
+        from mt_metadata.transfer_functions.processing.aurora.band import (
             FrequencyBands,
         )
